Enc and dec password

+++ starter/cloudstorage/src/main/java/com/udacity/jwdnd/course1/cloudstorage/controller/NoteController.java
@@ -1,12 +1,13 @@
 package com.udacity.jwdnd.course1.cloudstorage.controller;
 
 import com.udacity.jwdnd.course1.cloudstorage.formdata.NoteData;
-import com.udacity.jwdnd.course1.cloudstorage.mappers.CredentialData;
+import com.udacity.jwdnd.course1.cloudstorage.formdata.CredentialData;
 import com.udacity.jwdnd.course1.cloudstorage.model.Note;
 import com.udacity.jwdnd.course1.cloudstorage.model.User;
 import com.udacity.jwdnd.course1.cloudstorage.services.NoteService;
 import com.udacity.jwdnd.course1.cloudstorage.services.UserService;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -29,14 +30,18 @@ public class NoteController {
             Model model,
             @ModelAttribute("credentialData") CredentialData credentialData
     ) {
-        User user = userService.getUser(authentication.getName());
-        if (user == null) {
+        if (authentication != null) {
+            User user = userService.getUser(authentication.getName());
+            if (user == null) {
+                return "redirect:/login";
+            } else {
+                model.addAttribute("notes", noteService.getNotesByUserId(user.getUserId()));
+                model.addAttribute("isNote", true);
+                return "home";
+            }
+        } else {
             return "redirect:/login";
         }
-
-        model.addAttribute("notes", noteService.getNotesByUserId(user.getUserId()));
-        model.addAttribute("isNote", true);
-        return "home";
     }
 
     @PostMapping("/note")
@@ -45,22 +50,28 @@ public class NoteController {
             Authentication authentication, Model model,
             @ModelAttribute("credentialData") CredentialData credentialData
             ) {
-        User user = userService.getUser(authentication.getName());
-        if (user == null) {
+        if (authentication != null) {
+            User user = userService.getUser(authentication.getName());
+            if (user == null) {
+                SecurityContextHolder.clearContext();
+                SecurityContextHolder.getContext().setAuthentication(null);
+                return "redirect:/login";
+            } else {
+                Note note = new Note(noteData.getNoteTitle(), noteData.getNoteDescription(), String.valueOf(user.getUserId()));
+                if (noteData.getNoteId().isEmpty()) {
+                    noteService.insertNote(note);
+                } else  {
+                    note.setNoteId(Integer.parseInt(noteData.getNoteId()));
+                    noteService.updateNote(note);
+                }
+
+                model.addAttribute("notes", noteService.getNotesByUserId(user.getUserId()));
+                model.addAttribute("isNote", true);
+                return "home";
+            }
+        } else {
             return "redirect:/login";
         }
-
-        Note note = new Note(noteData.getNoteTitle(), noteData.getNoteDescription(), String.valueOf(user.getUserId()));
-        if (noteData.getNoteId().isEmpty()) {
-            noteService.insertNote(note);
-        } else  {
-            note.setNoteId(Integer.parseInt(noteData.getNoteId()));
-            noteService.updateNote(note);
-        }
-
-        model.addAttribute("notes", noteService.getNotesByUserId(user.getUserId()));
-        model.addAttribute("isNote", true);
-        return "home";
     }
 
     @GetMapping("/note/delete/{noteId}")
@@ -71,15 +82,21 @@ public class NoteController {
             Model model,
             @ModelAttribute("credentialData") CredentialData credentialData
     ) {
-        User user = userService.getUser(authentication.getName());
-        if (user == null) {
+        if (authentication != null) {
+            User user = userService.getUser(authentication.getName());
+            if (user == null) {
+                SecurityContextHolder.clearContext();
+                SecurityContextHolder.getContext().setAuthentication(null);
+                return "redirect:/login";
+            } else {
+                noteService.deleteNoteById(noteId);
+                model.addAttribute("notes", noteService.getNotesByUserId(user.getUserId()));
+                model.addAttribute("isNote", true);
+                return "home";
+            }
+        } else {
             return "redirect:/login";
         }
-
-        noteService.deleteNoteById(noteId);
-        model.addAttribute("notes", noteService.getNotesByUserId(user.getUserId()));
-        model.addAttribute("isNote", true);
-        return "home";
     }
 
 }
